src/requirements.txt
beautifulsoup4==4.12.3
dependency-injector==4.41.0
fastapi==0.110.1
google-cloud-logging==3.6.0
pandas==2.1.1
pydantic==2.6.4
uvicorn[standard]==0.29.0
websocket-client==1.6.3